Add files via upload
During my internship at the National Physical Laboratory (NPL), I contributed to a cutting-edge project to develop a real-time data acquisition and visualization system. This system plays a critical role in monitoring the process of cooling cesium atoms using an ionizer, which is essential for operating and maintaining atomic clocks.